requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
py==1.11.0
pytest==7.0.1
requests==2.32.3
tomli==2.0.2
urllib3==2.2.3
